Add files via upload
This is a part of Dataset Credit Card Fraud which you can find here:
https://www.kaggle.com/mlg-ulb/creditcardfraud

This task was to create a Classifier that can detect Fraud transactions on our bank card. This is extremely important task due to money laundring problem.
Due to small amount of Data about Fraud transactions key objective of this task was to use Undersampling techniques and OverSampling Technques.